requirements.txt
kafka-python==1.4.7
pymongo==4.3.3
paho-mqtt==1.6.1